+++ Source/APU/Channel.h
@@ -35,10 +35,7 @@ public:
 	CExChannel(CMixer *pMixer, uint8 Chip, uint8 ID) :
 		m_pMixer(pMixer),
 		m_iChip(Chip),
-		m_iChanId(ID),
-		m_iTime(0),
-		m_iLastValueLeft(0),		// // //
-		m_iLastValueRight(0)
+		m_iChanId(ID)		// // //
 	{
 	}
 
@@ -52,37 +49,19 @@ public:
 
 protected:
 	inline void Mix(int32 Value) {
-		MixLeft(Value);		// // //
-		MixRight(Value);
+		m_pMixer->AddValue(m_iChanId, m_iChip, Value, Value, m_iTime);		// // //
 	}
 	inline void Mix(int32 Left, int32 Right) {
-		MixLeft(Left);
-		MixRight(Right);
-	}
-
-private:
-	inline void MixSingle(int32 Value, int32 &Last, bool Right) {		// // //
-		int32 Delta = Value - Last;
-		if (Delta)
-			m_pMixer->AddValue(m_iChanId, m_iChip, Delta, Value, m_iTime, Right);
-		Last = Value;
-	}
-	inline void MixLeft(int32 Value) {		// // //
-		MixSingle(Value, m_iLastValueLeft, false);
-	}
-	inline void MixRight(int32 Value) {		// // //
-		MixSingle(Value, m_iLastValueRight, true);
+		m_pMixer->AddValue(m_iChanId, m_iChip, Left, Right, m_iTime);		// // //
 	}
 
 protected:
 	CMixer	*m_pMixer;
 	const CVGMWriterBase *m_pVGMWriter = nullptr;
 
-	uint32	m_iTime;			// Cycle counter, resets every new frame
+	uint32	m_iTime = 0;			// Cycle counter, resets every new frame
 	uint8	m_iChip;
 	uint8	m_iChanId;
-	int32	m_iLastValueLeft;		// Last value sent to mixer
-	int32	m_iLastValueRight;		// // // stereo support
 };
 
 #endif /* CHANNEL_H */